arreglo de un error de consola y reformat del codigo de movimientos
