+++ src/dominio/lista-movimientos.mjs
@@ -5,12 +5,19 @@ export default class ListaMovimientos {
   }
 
   agregar(movimiento) {
-    let movimientosAgregados = this.movimientos.some(m => m.nombre == movimiento.nombre);
+    let movimientosAgregados = this.movimientos.some(m => m.nombreMovimiento == movimiento.nombreMovimiento);
     if (!movimientosAgregados) {
       this.movimientos.push(movimiento);
     } else {
-      throw new Error(`No se pudo agregar. ${movimiento.nombre} ya se encuentra en la lista.`);
+      throw new Error(`No se pudo agregar. ${movimiento.nombreMovimiento} ya se encuentra en la lista.`);
     }
+  }
+  eliminar(movimiento) {
+    this.setMovimientos(this.movimientos.filter(c => c.nombreMovimiento !== categoria.nombreMovimiento));
+  }
+
+  setMovimientos(movimientos) {
+    this.movimientos = movimientos;
   }
 
   getMovimientos() {
