+++ src/dominio/movimiento.mjs
@@ -1,7 +1,7 @@
 export default class Movimiento {
 
-  constructor(nombre, tipo, fecha, valor) {
-    this.nombre = nombre;
+  constructor(nombreMovimiento, tipo, fecha, valor) {
+    this.nombre = nombreMovimiento;
     this.tipo = tipo;
     this.fecha = fecha;
     this.valor = valor;
