+++ src/interfaz/index.html
@@ -69,7 +69,7 @@
           </div>
           <div class="mdc-layout-grid__inner">
               <div class="mdc-layout-grid__cell--span-6">
-                  <ul class="mdc-list">
+                  <ul class="mdc-list movimientos__columna--lista" id="listar-movimientos">
                       <li class="mdc-list-item" tabindex="0">
                           <span class="mdc-list-item__ripple"></span>
                           <div>
@@ -89,16 +89,10 @@
                       </li>
                   </ul>
               </div>
-              <div class="mdc-layout-grid__cell--span-6">aca va el form</div>
-          </div>
-          <div class="mdc-layout-grid__inner">
-              <div class="mdc-layout-grid__cell--span-6">
-                  <ul class="movimientos__columna--lista" id="listar-movimientos"></ul>
-              </div>
               <div class="mdc-layout-grid__cell--span-6">
                   <h2>Añadir nuevo movimiento</h2>
 
-                  <label class="mdc-text-field mdc-text-field--outlined" id="nombre">
+                  <label class="mdc-text-field mdc-text-field--outlined" id="nombreMovimiento">
                         <span class="mdc-notched-outline">
                           <span class="mdc-notched-outline__leading"></span>
                           <span class="mdc-notched-outline__notch">
@@ -106,7 +100,7 @@
                           </span>
                           <span class="mdc-notched-outline__trailing"></span>
                         </span>
-                        <input type="text" class="mdc-text-field__input" aria-labelledby="nombre">
+                      <input type="text" class="mdc-text-field__input" aria-labelledby="nombreMovimiento">
                   </label>
 
                   <label class="mdc-text-field mdc-text-field--outlined" id="valor">
@@ -117,7 +111,7 @@
                           </span>
                           <span class="mdc-notched-outline__trailing"></span>
                         </span>
-                        <input type="number" class="mdc-text-field__input" aria-labelledby="valor">
+                      <input type="number" class="mdc-text-field__input" aria-labelledby="valor">
                   </label>
 
                   <label class="mdc-text-field mdc-text-field--outlined" id="fecha">
@@ -128,7 +122,7 @@
                           </span>
                           <span class="mdc-notched-outline__trailing"></span>
                         </span>
-                        <input type="date" class="mdc-text-field__input" aria-labelledby="fecha">
+                      <input type="date" class="mdc-text-field__input" aria-labelledby="fecha">
                   </label>
 
                   <div id="select-tipo" class="mdc-select mdc-select--outlined">
@@ -140,7 +134,7 @@
                         </span>
                         <span class="mdc-notched-outline__trailing"></span>
                       </span>
-                      <span class="mdc-select__selected-text-container">
+                          <span class="mdc-select__selected-text-container">
                         <span id="demo-selected-text-tipo" class="mdc-select__selected-text"></span>
                       </span>
                           <span class="mdc-select__dropdown-icon">
@@ -173,17 +167,20 @@
                   </div>
 
                   <div class="button">
-                      <button class="mdc-button mdc-button--raised" id="addButton">
+                      <button class="mdc-button mdc-button--raised" id="agregarMovimientoBtn">
                           <span class="mdc-button__label">AGREGAR</span>
                       </button>
                   </div>
               </div>
           </div>
       </section>
+
       <section class="content sample-content--hidden">
+          <div class="mdc-layout-grid__inner">
+              <h1>Categorias</h1>
+          </div>
           <div class="mdc-layout-grid__inner categoria">
               <div class="mdc-layout-grid__cell--span-6 categoria__columna">
-                  <h1>Categorias</h1>
                   <ul class="categoria__columna--lista" id="listar-categorias"></ul>
               </div>
               <div class="mdc-layout-grid__cell--span-6 categoria__columna">
